// src/enums/abc/OperatorAbc.ts
enum OperatorAbc {
  SHARP = '^',
  FLAT = '_',
  DOUBLE_SHARP = '^^',
  DOUBLE_FLAT = '__',
  NATURAL = '=',
  UPPER_OCTAVE = "'",
  LOWER_OCTAVE = ',',
  START_CHORD = '[',
  END_CHORD = ']',
  START_TUPLET = '(',
  END_TUPLET = ')',
  START_REPEAT = '|:',
  END_REPEAT = ':|',
}

export default OperatorAbc